cloud_function/requirements.txt
functions-framework==3.*
langchain
pypdf
PyPDF2
docx2txt
chromadb
google-cloud-storage
google-cloud-aiplatform
langchain-community
python-docx
langchain_google_genai
google-generativeai
flask
python-docx